+++ CHANGELOG.md
@@ -62,13 +62,13 @@ _**Note**: This version is a complete re-write of the library based on a custom 
  - Show error when config file does not exist !86
  - ~~Doxygen documentation is now deployed to Netlify !73~~ (reverted on !176)
 ### Changed
+ - Update docusaurus documentation to latest changes !176 !178 !88
  - Math parsers now can only take move only functions !187
  - Reduce operations now receive one less argument !184
  - Reorganize files and add `DiffusionReaction` namespace !183
  - Use grid leaf view instead of level 0 grid view for grid cell data !177
  - Author email address !181
  - Improve support for Wasm binaries !170 !182
- - Update docusaurus documentation to latest changes !176 !178
  - Move configuration options to a JSON file and use schema validator on it !174
  - Use Debian Bookworm in the CI !171
  - Conditinally test possion config if ExprTk is available !168